Made CLI documentation more accurate.
The "atom" input at the command line is meant to mean the orbital.
For 1s-only simulations this is just like counting atoms but
when you have multiple core states per atom (1s, 2s, 2p, etc) it
is very much not. It's easier just to run pynex with atom set to 1
the first time to figure out how many core projectors are actually
in the eels_mat file. Since the special core-hole atom should be at
the END of the orbitals list you can just count it backwards from
that number.