Modification to lowest transition calculation.
Extensive changes to how the lowest state is calculated based
on the real hole charge spin difference. Also, nelectrons
is no longer read from .bands and must be supplied prior
to init (as with hole_charge). This is because CASTEP doesn't
report the true spin populations in the .bands file.